+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * 2.6.6
     - added debug message for invalid coverage
+    - fixed snapshot without debug file
 
 * 2.6.5
     - fixed sourcemap issue for multiple lines of original text